Changed several function names
This was done in response to inconsistencies that presented themselves
during training. Here are the fixes in this commit:

* `MAXIS_check_function`: is now `MAXIS_check(end_script)`, and includes
a True/False for whether or not the script should end, or just loop
waiting for the MAXIS screen to be found again. As
`MAXIS_check_function` always stopscript-ed here, the original function
passes the variable `True` to the new function.
* `write_editbox_in_case_note`: is now
`write_bullet_and_variable_in_case_note`. I know it's a longer name, but
it's far more descriptive (I get asked "can you put dropdowns in this
function?" a lot). Also, I removed the "length of indent" parameter from
the new one (the old one still asks for it, it just isn't passed
anywhere anymore). It will always default to 6 now.
* `write_new_line_in_case_note`: is now `write_variable_in_case_note`.
All other functionality remains the same.
* `write_new_line_in_SPEC_MEMO`: is now `write_variable_in_SPEC_MEMO`.
All other functionality remains the same.
* `write_TIKL_function`: is now `write_variable_in_TIKL`. All other
functionality remains the same.

In addition to the above changes, I also cleaned up comments in a few
places here and there.

I'll email out scriptwriters, and set forth a plan to remove the old
functions from existing scripts. After that, I'll leave them intact
here, but with a MsgBox explaining how a scriptwriter should enhance
them. A few months after that, I'll simply remove them.

Progress!!!